fix(e2e): one helper for smart-table column filters, and use it in remove-user
remove-user got past the Add New dialog and then failed verifying the user it
had just created. The snapshot showed why: the Full Name filter box held the
name, but the grid still read "1 - 10 of 14 Items" and listed everybody — the
filter had not been applied at all, and the new user was on page 2.

The filter cell is `<input [value]="query" (change) (keyup)>`. It never listens
for 'input', which is the only event .fill() dispatches. Three page objects had
independently hit this and two had just been fixed in place, so fold the working
version into util as applySmartTableFilter (fill + dispatch 'change' + read the
value back, since the debounced refetch writes `query` back into [value]) and
route DeleteOrganization, OrganizationPublicPage and RemoveUser through it.

Verified after the refactor: remove-user, edit-user, delete-organization,
organization-public-page and add-user all pass.
