+++ apps/gauzy-e2e/tests/support/pages/RemoveUser.po.ts
@@ -1,4 +1,5 @@
 import {
+	applySmartTableFilter,
 	verifyElementIsVisible,
 	verifyElementIsVisibleByIndex,
 	waitElementToHide,
@@ -23,18 +24,15 @@ export const clickGridButton = async () => {
 // Filter the users grid by Full Name so the just-added user is the only data row on page 1. The shared
 // serial DB accumulates users from earlier specs (seed admin + faker employees from addEmployee), so the
 // grid paginates and a freshly-added user can land on page 2 — invisible to a filter-by-text verify or a
-// row click. Typing the unique name into the smart-table Full Name filter narrows the grid to that record.
-// Mirrors the proven ManageEmployees.searchEmployeeByName pattern.
+// row click.
+//
+// This used to be a bare .fill(), which sets the box's text but dispatches only 'input' — an event the
+// smart-table InputFilterComponent does not listen for. The failing run's snapshot proves it: the Full
+// Name box held the new user's full name while the grid still showed "1 - 10 of 14 Items" with every
+// user on it, and the new user sat on page 2, never rendered. applySmartTableFilter dispatches the
+// 'change' the component actually subscribes to and reads the value back.
 export const filterByName = async (name: string) => {
-	const page = getPage();
-	await waitForSpinnerGone();
-	await page.waitForLoadState('networkidle').catch(() => {});
-	const filter = page.locator(RemoveUserPage.nameFilterInputCss).first();
-	await filter.fill(String(name)).catch(() => {});
-	// smart-table filtering is debounced; let the grid re-render before verifying/selecting.
-	await page.waitForTimeout(2000);
-	await waitForSpinnerGone();
-	await page.waitForLoadState('networkidle').catch(() => {});
+	await applySmartTableFilter(RemoveUserPage.nameFilterInputCss, name);
 };
 
 export const tableBodyExists = async () => {
